requirements.txt
streamlit~=1.35.0
requests~=2.25.1
python-dotenv~=0.17.1
